requirements.txt
PyQt6>=6.6.1
pypdf>=4.3.1
pypdfium2>=4.28.0
Pillow>=10.4.0
